fastApi/app/requirements.txt
anyio==3.6.1
asgiref
click
dnspython==2.2.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
motor
mysql-connector-python==8.0.32
pydantic==1.9.1
PyMySQL==1.0.2
python-dotenv==0.20.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions
uvicorn
requests
alembic
sqlalchemy
sqlModel
riotwatcher
pandas
numpy
scikit-learn
python-decouple